requirement.txt
autoattack @ git+https://github.com/fra31/auto-attack.git@a39220048b3c9f2cca9a4d3a54604793c68eca7e
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
eagerpy==0.30.0
filelock==3.18.0
fonttools==4.58.0
foolbox==3.3.4
fsspec==2025.5.0
gdown==5.1.0
geotorch==0.3.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.2
huggingface-hub==0.32.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
robustbench @ git+https://github.com/RobustBench/robustbench.git@78fcc9e48a07a861268f295a777b975f25155964
safetensors==0.5.3
scipy==1.15.3
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sympy==1.14.0
timm==1.0.15
torch==2.7.0
torchdiffeq==0.2.5
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
